improve bottom tab bar performance
